+++ packages/client/src/views/SysAdmin/Config/Forms/Home/DraftsTab.tsx
@@ -25,7 +25,8 @@ import {
   draftStatusMessages
 } from '@client/i18n/messages/views/formConfig'
 import { LinkButton } from '@opencrvs/components/lib/buttons'
-import { ToggleMenu, Pill } from '@opencrvs/components/lib/interface'
+import { ToggleMenu } from '@opencrvs/components/lib/interface'
+import { Pill } from '@opencrvs/components/lib/Pill'
 import { VerticalThreeDots } from '@opencrvs/components/lib/icons'
 import { goToFormConfigWizard } from '@client/navigation'
 import { IFormDraft } from '@client/forms/configuration/formDrafts/utils'
