+++ packages/components/src/interface/index.ts
@@ -30,4 +30,3 @@ export * from './ListTable'
 export * from './LocationSearch/LocationSearch'
 export * from './ErrorToastNotification'
 export * from './TableView'
-export * from './Pill'
